St1ckyweb README.md
This is barely an introduction to myself, but I will add onto it later once I learn more about programming languages. Just trying stuff out, I really hope to be in the cyber-security field once I'm older. I'm willing to do anything to achieve this goal. Blood, sweat, and tears are on the ready to be excreted from my body, just to name a few.